+++ prog/dftb+/lib_dftbplus/initprogram.F90
@@ -69,9 +69,11 @@ module dftbp_initprogram
   use dftbp_linresp_module
   use dftbp_stress
   use dftbp_orbitalequiv
+  use dftbp_orbitals
   use dftbp_commontypes
   use dftbp_sorting, only : heap_sort
   use dftbp_linkedlist
+  use dftbp_wrappedintr
   use dftbp_xlbomd_module
   use dftbp_etemp, only : Fermi
 #:if WITH_SOCKETS
@@ -1065,6 +1067,8 @@ contains
     !> Nr. of buffered Cholesky-decompositions
     integer :: nBufferedCholesky
 
+    character(sc), allocatable :: shellNamesTmp(:)
+
     @:ASSERT(input%tInitialized)
 
     write(stdOut, "(/, A)") "Starting initialization..."
@@ -1200,8 +1204,6 @@ contains
     atomEigVal(:,:) = input%slako%skSelf(1:orb%mShell, :)
 
     @:ASSERT(size(input%slako%skOcc, dim=1) >= orb%mShell)
-    allocate(referenceN0(orb%mShell, nType))
-    referenceN0(:,:) = input%slako%skOcc(1:orb%mShell, :)
     @:ASSERT(size(input%slako%mass) == nType)
     allocate(speciesMass(nType))
     speciesMass(:) = input%slako%mass(:)
@@ -1368,8 +1370,8 @@ contains
       ! Longest cut-off including the softening part of gamma
       mCutOff = max(mCutOff, sccCalc%getCutOff())
 
-      if (input%ctrl%t3rd .and. input%ctrl%tOrbResolved) then
-        call error("Onsite third order DFTB only compatible with orbital non resolved SCC")
+      if (input%ctrl%t3rd .and. input%ctrl%tShellResolved) then
+        call error("Onsite third order DFTB only compatible with shell non-resolved SCC")
       end if
 
       ! Initialize full 3rd order module
@@ -1383,7 +1385,7 @@ contains
         allocate(thirdInp%damped(nType))
         thirdInp%damped(:) = tDampedShort
         thirdInp%dampExp = input%ctrl%dampExp
-        thirdInp%shellResolved = input%ctrl%tOrbResolved
+        thirdInp%shellResolved = input%ctrl%tShellResolved
         allocate(thirdOrd)
         call ThirdOrder_init(thirdOrd, thirdInp)
         mCutOff = max(mCutOff, thirdOrd%getCutOff())
@@ -1408,6 +1410,7 @@ contains
       call THalogenX_init(halogenXCorrection, species0, speciesName)
     end if
 
+    allocate(referenceN0(orb%mShell, nType))
     allocate(mass(nAtom))
     mass = speciesMass(species0)
     if (allocated(input%ctrl%masses)) then
@@ -1448,28 +1451,6 @@ contains
     else
       allocate(nEl(nSpin))
       allocate(Ef(nSpin))
-    end if
-
-    nEl0 = 0.0_dp
-    do ii = 1, nAtom
-      nEl0 = nEl0 + sum(input%slako%skOcc(1:orb%nShell(species0(ii)), species0(ii)))
-    end do
-    nEl(:) = 0.0_dp
-    if (nSpin == 1 .or. nSpin == 4) then
-      nEl(1) = nEl0 - input%ctrl%nrChrg
-      if(ceiling(nEl(1)) > 2.0_dp*nOrb) then
-        call error("More electrons than basis functions!")
-      end if
-    else
-      nEl(1) = 0.5_dp * (nEl0 - input%ctrl%nrChrg + input%ctrl%nrSpinPol)
-      nEl(2) = 0.5_dp * (nEl0 - input%ctrl%nrChrg - input%ctrl%nrSpinPol)
-      if (any(ceiling(nEl(:)) > nOrb)) then
-        call error("More electrons than basis functions!")
-      end if
-    end if
-
-    if (.not.all(nEl(:) >= 0.0_dp)) then
-      call error("Less than 0 electrons!")
     end if
 
     iDistribFn = input%ctrl%iDistribFn
@@ -1651,6 +1632,56 @@ contains
 
     tPrintForces = input%ctrl%tPrintForces
     tForces = input%ctrl%tForces .or. tPrintForces
+    tLinResp = input%ctrl%lrespini%tInit
+
+    referenceN0(:,:) = input%slako%skOcc(1:orb%mShell, :)
+
+    ! Allocate charge arrays
+    if (tMulliken) then ! automatically true if tSccCalc
+      allocate(q0(orb%mOrb, nAtom, nSpin))
+      q0(:,:,:) = 0.0_dp
+      allocate(qShell0(orb%mShell, nAtom))
+      qShell0(:,:) = 0.0_dp
+    else
+      allocate(q0(0,0,0))
+      allocate(qShell0(0,0))
+    end if
+
+    ! Initialize reference neutral atoms.
+    if (tLinResp.and.allocated(input%ctrl%customOccAtoms)) then 
+       call error("Custom occupation not compatible with linear response")
+    end if     
+    if (tMulliken) then
+      if (allocated(input%ctrl%customOccAtoms)) then
+        call applyCustomReferenceOccupations(input%ctrl%customOccAtoms, &
+            & input%ctrl%customOccFillings, species0, orb, referenceN0, q0)
+      else   
+        call initQFromShellChrg(q0, referenceN0, species0, orb)
+      end if  
+    end if
+
+    nEl0 = sum(q0(:,:,1))
+    if (abs(nEl0 - nint(nEl0)) < elecTolMax) then
+      nEl0 = nint(nEl0)
+    end if
+    nEl(:) = 0.0_dp
+    if (nSpin == 1 .or. nSpin == 4) then
+      nEl(1) = nEl0 - input%ctrl%nrChrg
+      if(ceiling(nEl(1)) > 2.0_dp*nOrb) then
+        call error("More electrons than basis functions!")
+      end if
+    else
+      nEl(1) = 0.5_dp * (nEl0 - input%ctrl%nrChrg + input%ctrl%nrSpinPol)
+      nEl(2) = 0.5_dp * (nEl0 - input%ctrl%nrChrg - input%ctrl%nrSpinPol)
+      if (any(ceiling(nEl(:)) > nOrb)) then
+        call error("More electrons than basis functions!")
+      end if
+    end if
+  
+    if (.not.all(nEl(:) >= 0.0_dp)) then
+      call error("Less than 0 electrons!")
+    end if
+
     if (tForces) then
       tCasidaForces = input%ctrl%tCasidaForces
     else
@@ -1694,6 +1725,11 @@ contains
 
     if (electronicSolver%isElsiSolver) then
       @:ASSERT(parallelKS%nLocalKS == 1)
+
+      if (omp_get_max_threads() > 1) then
+        call error("The ELSI-solvers should not be run with multiple threads. Set the&
+            & environment variable OMP_NUM_THREADS to 1 in order to disable multi-threading.")
+      end if
 
       ! Would be using the ELSI matrix writing mechanism, so set this as always false
       tWriteHS = .false.
@@ -1904,19 +1940,15 @@ contains
       call TElStatPotentials_init(esp, input%ctrl%elStatPotentialsInp, tEField .or. tExtChrg)
     end if
 
-    tLocalise = input%ctrl%tLocalise
-    if (tLocalise .and. (nSpin > 2 .or. t2Component)) then
-      call error("Localisation of electronic states currently unsupported for non-collinear and&
-          & spin orbit calculations")
-    end if
     if (allocated(input%ctrl%pipekMezeyInp)) then
       allocate(pipekMezey)
       call initialise(pipekMezey, input%ctrl%pipekMezeyInp)
     end if
     tLocalise = allocated(pipekMezey)
-
-    ! Linear response
-    tLinResp = input%ctrl%lrespini%tInit
+    if (tLocalise .and. (nSpin > 2 .or. t2Component)) then
+      call error("Localisation of electronic states currently unsupported for non-collinear and&
+          & spin orbit calculations")
+    end if
 
     if (tLinResp) then
 
@@ -1939,8 +1971,8 @@ contains
         call error("Linear response does not support spin orbit coupling at the moment.")
       elseif (tDFTBU) then
         call error("Linear response does not support LDA+U yet")
-      elseif (input%ctrl%tOrbResolved) then
-        call error("Linear response does not support orbital resolved scc yet")
+      elseif (input%ctrl%tShellResolved) then
+        call error("Linear response does not support shell resolved scc yet")
       end if
       if (tempElec > 0.0_dp .and. tCasidaForces) then
         write(tmpStr, "(A,E12.4,A)")"Excited state forces are not implemented yet for fractional&
@@ -2133,18 +2165,6 @@ contains
       EfieldPhase = 0
     end if
 
-    ! Allocate charge arrays
-    if (tMulliken) then ! automatically true if tSccCalc
-      allocate(q0(orb%mOrb, nAtom, nSpin))
-      q0(:,:,:) = 0.0_dp
-
-      allocate(qShell0(orb%mShell, nAtom))
-      qShell0(:,:) = 0.0_dp
-    else
-      allocate(q0(0,0,0))
-      allocate(qShell0(0,0))
-    end if
-
     allocate(qInput(orb%mOrb, nAtom, nSpin))
     allocate(qOutput(orb%mOrb, nAtom, nSpin))
     qInput(:,:,:) = 0.0_dp
@@ -2175,11 +2195,6 @@ contains
       qOutRed = 0.0_dp
     end if
 
-    ! Initialize Mulliken charges
-    if (tMulliken .or. tLinResp) then
-      call initQFromShellChrg(q0, referenceN0, species0, orb)
-    end if
-
     tReadChrg = input%ctrl%tReadChrg
 
     tReadShifts = input%ctrl%tReadShifts
@@ -2190,17 +2205,13 @@ contains
 
     tWriteChrgAscii = input%ctrl%tWriteChrgAscii
 
-    if (tReadChrg) then
-      tSkipChrgChecksum = input%ctrl%tSkipChrgChecksum .or. tNegf
-    else
-      tSkipChrgChecksum = .false.
-    end if
+    tSkipChrgChecksum = input%ctrl%tSkipChrgChecksum .or. tNegf
 
     if (tSccCalc) then
       do iAt = 1, nAtom
         iSp = species0(iAt)
         do iSh = 1, orb%nShell(iSp)
-          qShell0 (iSh,iAt) = sum(q0(orb%posShell(iSh,iSp):orb%posShell(iSh+1,iSp)-1,iAt,1))
+          qShell0(iSh,iAt) = sum(q0(orb%posShell(iSh,iSp):orb%posShell(iSh+1,iSp)-1,iAt,1))
         end do
       end do
       if (tReadChrg) then
@@ -2215,7 +2226,7 @@ contains
             end if
           else
             if (tImHam) then
-              if (tFixEf .or. input%ctrl%tSkipChrgChecksum) then
+              if (tFixEf .or. tSkipChrgChecksum) then
                 ! do not check charge or magnetisation from file
                 call initQFromFile(qInput, fCharges, input%ctrl%tReadChrgAscii, orb,&
                     & qBlock=qBlockIn,qiBlock=qiBlockIn)
@@ -2495,7 +2506,7 @@ contains
               tmpir1 = 0
               ind = 1
               do iAt = 1, nAtomRegion
-                tmpir1(ind) = denseDesc%iAtomStart(iAt) + iOrb - 1
+                tmpir1(ind) = denseDesc%iAtomStart(iAtomRegion(iAt)) + iOrb - 1
                 ind = ind + 1
               end do
               call append(iOrbRegion, tmpir1)
@@ -2664,7 +2675,7 @@ contains
       write(stdOut, "(A,':',T30,A)") "Self consistent charges", "Yes"
       write(stdOut, "(A,':',T30,E14.6)") "SCC-tolerance", sccTol
       write(stdOut, "(A,':',T30,I14)") "Max. scc iterations", maxSccIter
-      if (input%ctrl%tOrbResolved) then
+      if (input%ctrl%tShellResolved) then
          write(stdOut, "(A,':',T30,A)") "Shell resolved Hubbard", "Yes"
       else
          write(stdOut, "(A,':',T30,A)") "Shell resolved Hubbard", "No"
@@ -2741,18 +2752,6 @@ contains
       end if
     end if
 
-    if (tForces) then
-      select case (forceType)
-      case(forceTypes%orig)
-        strTmp = "Traditional"
-      case(forceTypes%dynamicT0)
-        strTmp = "Dynamics, zero electronic temp."
-      case(forceTypes%dynamicTFinite)
-        strTmp = "Dynamics, finite electronic temp."
-      end select
-      write(stdOut, "(A,':',T30,A)") "Force evaluation method", trim(strTmp)
-    end if
-
     tFirst = .true.
     if (allocated(conAtom)) then
       do ii = 1, nAtom
@@ -2793,6 +2792,7 @@ contains
     end if
 
     do iSp = 1, nType
+      call getShellNames(iSp, orb, shellNamesTmp)
       if (iSp == 1) then
         write (strTmp, "(A,':')") "Included shells"
       else
@@ -2800,13 +2800,21 @@ contains
       end if
       do jj = 1, orb%nShell(iSp)
         if (jj == 1) then
-          strTmp2 = trim(orbitalNames(orb%angShell(jj, iSp) + 1))
+          strTmp2 = trim(shellNamesTmp(jj))
         else
-          strTmp2 = trim(strTmp2) // ", " // trim(orbitalNames(orb%angShell(jj, iSp) + 1))
+          strTmp2 = trim(strTmp2) // ", " // trim(shellNamesTmp(jj))
         end if
       end do
       write(stdOut, "(A,T29,A2,':  ',A)") trim(strTmp), trim(speciesName(iSp)), trim(strTmp2)
+      deallocate(shellNamesTmp)
     end do
+
+    if (tMulliken) then
+      if (allocated(input%ctrl%customOccAtoms)) then
+        call printCustomReferenceOccupations(orb, input%geom%species, &
+            & input%ctrl%customOccAtoms, input%ctrl%customOccFillings)
+      end if
+    end if
 
     if (tPeriodic) then
       do ii = 1, nKPoint
@@ -2862,7 +2870,7 @@ contains
               write(strTmp, "(A)") ""
             end if
             write(stdOut, "(A,T30,A2,2X,I1,'(',A1,'): ',E14.6)") trim(strTmp), speciesName(iSp),&
-                & jj, orbitalNames(orb%angShell(jj, iSp)+1), hubbU(jj, iSp)
+                & jj, shellNames(orb%angShell(jj, iSp)+1), hubbU(jj, iSp)
           end do
         end do
       end if
@@ -2880,8 +2888,8 @@ contains
               write(strTmp, "(A)") ""
             end if
             write(stdOut, "(A,T30,A2,2X,I1,'(',A1,')-',I1,'(',A1,'): ',E14.6)")trim(strTmp),&
-                & speciesName(iSp), jj, orbitalNames(orb%angShell(jj, iSp)+1), kk,&
-                & orbitalNames(orb%angShell(kk, iSp)+1), spinW(kk, jj, iSp)
+                & speciesName(iSp), jj, shellNames(orb%angShell(jj, iSp)+1), kk,&
+                & shellNames(orb%angShell(kk, iSp)+1), spinW(kk, jj, iSp)
           end do
         end do
       end do
@@ -2901,7 +2909,7 @@ contains
             write(strTmp, "(A)") ""
           end if
           write(stdOut, "(A,T30,A2,2X,I1,'(',A1,'): ',E14.6)")trim(strTmp), speciesName(iSp),&
-                & jj, orbitalNames(orb%angShell(jj, iSp)+1), xi(jj, iSp)
+                & jj, shellNames(orb%angShell(jj, iSp)+1), xi(jj, iSp)
           if (xi(jj, iSp) /= 0.0_dp .and. orb%angShell(jj, iSp) == 0) then
             call error("Program halt due to non-zero s-orbital spin-orbit coupling constant!")
           end if
@@ -2912,8 +2920,8 @@ contains
     if (tSccCalc) then
       if (t3rdFull) then
         write(stdOut, "(A,T30,A)") "Full 3rd order correction", "Yes"
-        if (input%ctrl%tOrbResolved) then
-          write(stdOut, "(A,T30,A)") "Orbital-resolved 3rd order", "Yes"
+        if (input%ctrl%tShellResolved) then
+          write(stdOut, "(A,T30,A)") "Shell-resolved 3rd order", "Yes"
           write(stdOut, "(A30)") "Shell-resolved Hubbard derivs:"
           write(stdOut, "(A)") "        s-shell   p-shell   d-shell   f-shell"
           do iSp = 1, nType
@@ -2946,15 +2954,17 @@ contains
     if (tPrintForces .and. .not. (tMD .or. tGeoOpt .or. tDerivs)) then
       write(stdOut, "(T30,A)") "Force calculation"
     end if
-    select case (forceType)
-    case(forceTypes%orig)
-      write(stdOut, "(A,T30,A)") "Force type", "original"
-    case(forceTypes%dynamicT0)
-      write(stdOut, "(A,T30,A)") "Force type", "erho with re-diagonalized eigenvalues"
-      write(stdOut, "(A,T30,A)") "Force type", "erho with DHD-product (T_elec = 0K)"
-    case(forceTypes%dynamicTFinite)
-      write(stdOut, "(A,T30,A)") "Force type", "erho with S^-1 H D (Te <> 0K)"
-    end select
+    if (tForces) then
+      select case (forceType)
+      case(forceTypes%orig)
+        write(stdOut, "(A,T30,A)") "Force type", "original"
+      case(forceTypes%dynamicT0)
+        write(stdOut, "(A,T30,A)") "Force type", "erho with re-diagonalized eigenvalues"
+        write(stdOut, "(A,T30,A)") "Force type", "erho with DHD-product (T_elec = 0K)"
+      case(forceTypes%dynamicTFinite)
+        write(stdOut, "(A,T30,A)") "Force type", "erho with S^-1 H D (Te <> 0K)"
+      end select
+    end if
     if (tPrintEigVecs) then
       write(stdOut, "(T30,A)") "Eigenvector printing"
     end if
@@ -3009,23 +3019,13 @@ contains
               end if
               write(stdOut, "(A,T30,A5,2X,I1,'(',A1,')-',I1,'(',A1,'): ',E14.6)")trim(strTmp),&
                   & trim(speciesName(iSp))//trim(strTmp2), jj,&
-                  & orbitalNames(orb%angShell(jj, iSp)+1), kk,&
-                  & orbitalNames(orb%angShell(kk, iSp)+1), onSiteElements(kk, jj, iSpin, iSp)
+                  & shellNames(orb%angShell(jj, iSp)+1), kk,&
+                  & shellNames(orb%angShell(kk, iSp)+1), onSiteElements(kk, jj, iSpin, iSp)
             end do
           end do
         end do
       end do
     end if
-
-    select case (forceType)
-    case(forceTypes%orig)
-      write(stdOut, "(A,T30,A)") "Force type", "original"
-    case(forceTypes%dynamicT0)
-      write(stdOut, "(A,T30,A)") "Force type", "erho with re-diagonalized eigenvalues"
-      write(stdOut, "(A,T30,A)") "Force type", "erho with DHD-product (T_elec = 0K)"
-    case(forceTypes%dynamicTFinite)
-      write(stdOut, "(A,T30,A)") "Force type", "erho with S^-1 H D (Te <> 0K)"
-    end select
 
     if (tSpinOrbit .and. tDFTBU .and. .not. tDualSpinOrbit)  then
       call error("Only dual spin orbit currently supported for orbital potentials")
@@ -3959,6 +3959,85 @@ contains
 
   end subroutine ensureSolverCompatibility
 
+  subroutine applyCustomReferenceOccupations(customOccAtoms, &
+      & customOccFillings, species, orb, referenceN0, q0)
+    type(WrappedInt1), allocatable, intent(in) :: customOccAtoms(:)
+    real(dp), intent(in) :: customOccFillings(:,:)
+    integer, intent(in) :: species(:)
+    type(TOrbitals), intent(in) :: orb
+    real(dp), intent(in) :: referenceN0(:,:)
+    real(dp), intent(inout) :: q0(:,:,:)
+    
+    integer :: nCustomBlock, iCustomBlock, iCustomAtom, nAtom, iAt, iSp
+    real(dp), allocatable :: refOcc(:,:)
+    
+    nAtom = size(species)
+    ! note that all arrays, referenceN0, customOccAtoms, refOcc 
+    ! are allocated to orb%mShell so assignments vecA(:,) = vecB(:,) work
+    allocate(refOcc(orb%mShell, nAtom))
+    ! initialize to referenceN0
+    do iAt = 1, nAtom
+      iSp = species(iAt)
+      refOcc(:, iAt) = referenceN0(:, iSp)
+    end do
+
+    ! override to customOccupation 
+    if (allocated(customOccAtoms)) then
+      nCustomBlock = size(customOccAtoms)
+      do iCustomBlock = 1, nCustomBlock
+        do iCustomAtom = 1, size(customOccAtoms(iCustomBlock)%data)
+          iAt =  customOccAtoms(iCustomBlock)%data(iCustomAtom)
+          refOcc(:, iAt) = customOccFillings(:,iCustomBlock)
+        end do
+      end do
+    end if
+
+    ! initialize q0 with right orbital order
+    call initQFromUsrChrg(q0, refOcc, species, orb)
+
+  end subroutine applyCustomReferenceOccupations
+  
+
+  subroutine printCustomReferenceOccupations(orb, species, customOccAtoms, &
+      & customOccFillings)
+    type(TOrbitals), intent(in) :: orb
+    integer, intent(in) :: species(:)
+    type(WrappedInt1), intent(in) :: customOccAtoms(:)
+    real(dp), intent(in) :: customOccFillings(:,:)
+
+    character(lc) :: formstr, outStr
+    integer :: nCustomBlock, iCustomBlock, iSp, nShell, nAtom, iSh
+    character(sc), allocatable :: shellnames(:)
+
+    nCustomBlock = size(customOccFillings)
+    if (nCustomBlock == 0) then
+      return
+    end if
+    write(stdout, "(A)") "Custom defined reference occupations:"
+    do iCustomBlock = 1, size(customOccAtoms)
+      nAtom = size(customOccAtoms(iCustomBlock)%data)
+      if (nAtom == 1) then
+        write(outStr, "(A)") "Atom:"
+      else
+        write(outStr, "(A)") "Atoms:"
+      end if
+      write(formstr, "(I0,A)") nAtom, "(I0,1X))"
+      write(stdout, "(A,T30,"//trim(formstr)//")") trim(outStr), customOccAtoms(iCustomBlock)%data
+      iSp = species(customOccAtoms(iCustomBlock)%data(1))
+      nShell = orb%nShell(iSp)
+      call getShellNames(iSp, orb, shellnames)
+      outStr = ""
+      do iSh = 1, nShell
+        if (iSh > 1) then
+          write(outStr,"(A,',')")trim(outStr)
+        end if
+        write(outStr,"(A,1X,A,F8.4)")trim(outStr), trim(shellnames(iSh)),&
+            & customOccFillings(iSh, iCustomBlock)
+      end do
+      write(stdout,"(A,T29,A)")"Fillings:",trim(outStr)
+      deallocate(shellnames)
+    end do
+  end subroutine printCustomReferenceOccupations
 
   !> Initialises SCC related parameters before geometry loop starts
   function getMinSccIters(tSccCalc, tDftbU, nSpin) result(minSccIter)