+++ prog/dftb+/lib_dftbplus/inputdata.F90
@@ -19,6 +19,7 @@ module dftbp_inputdata_module
   use dftbp_commontypes
   use dftbp_repcont
   use dftbp_linkedlist
+  use dftbp_wrappedintr
   use dftbp_elecsolvers, only : TElectronicSolverInp
   use dftbp_xlbomd_module
 #:if WITH_SOCKETS
@@ -87,7 +88,7 @@ module dftbp_inputdata_module
     logical :: tScc        = .false.
 
     !> l-shell resolved SCC
-    logical :: tOrbResolved = .false.
+    logical :: tShellResolved = .false.
 
     !> SCC tolerance
     real(dp) :: sccTol      = 0.0_dp
@@ -452,6 +453,10 @@ module dftbp_inputdata_module
 
     !> Maximal timing level to show in output
     integer :: timingLevel
+    
+    ! Custom occupations
+    type(WrappedInt1), allocatable :: customOccAtoms(:)
+    real(dp), allocatable :: customOccFillings(:,:)
 
   end type control