+++ prog/dftb+/lib_type/orbitals.F90
@@ -8,10 +8,12 @@
 !> Contains a type with basis information
 module dftbp_orbitals
   use dftbp_accuracy
+  use dftbp_message
+  use dftbp_constants, only : shellNames
   implicit none
   private
 
-  public :: TOrbitals
+  public :: TOrbitals, getShellNames, orbitalNames
 
 
   !> Contains information about the orbitals of the species/atoms in the system
@@ -46,5 +48,60 @@ module dftbp_orbitals
     !> Total number of orbitals in system.
     integer :: nOrb
   end type TOrbitals
+
+  !> Length of labels for atomic orbitals
+  integer, parameter :: lenOrbitalNames = 9
+
+  !> Names of the atomic orbitals in tesseral basis
+  !> general set for f orbitals (not cubic), see
+  !> http://winter.group.shef.ac.uk/orbitron/AOs/4f/equations.html
+  character(lenOrbitalNames), parameter :: orbitalNames(-3:3,0:3) = reshape([&
+      & '         ','         ','         ','         ','         ','         ','         ',&
+      & '         ','         ','y        ','z        ','x        ','         ','         ',&
+      & '         ','xy       ','yz       ','z2       ','xz       ','x2-y2    ','         ',&
+      & 'y(3x2-y2)','x2+y2+z2 ','yz2      ','z3       ','xz2      ','z(x2-y2) ','x(x2-3y2)'&
+      &], [7,4])
+
+contains
+
+  !> Builds a unique names for the atomic orbitals
+  !> Assign 's', 'p', 'd' to first occurring shells, then 's2', 'p2', ...
+  subroutine getShellNames(iSpecie, orb, shellNamesTmp)
+
+    !> atomic specie
+    integer, intent(in) :: iSpecie
+
+    !> orbital info
+    type(TOrbitals), intent(in) :: orb
+
+    !> output string naming the atomic orbital
+    character(sc), intent(out), allocatable :: shellNamesTmp(:)
+
+    integer :: ii, jj
+    integer, allocatable :: ind(:)
+    character(sc) :: sindx
+
+    !allocate(names(orb%nShell(iSpecie)))
+    allocate(shellNamesTmp(orb%nShell(iSpecie)))
+    allocate(ind(orb%nShell(iSpecie)))
+    ind = 1
+
+    do ii = 1, orb%nShell(iSpecie)
+      write(shellNamesTmp(ii), "(A)") shellNames(orb%angShell(ii, iSpecie) + 1)
+      if (any(shellNamesTmp(ii) == shellNamesTmp(1:ii-1))) then
+        ! at least one example of this shell already
+        ind(ii) = ind(ii) + 1
+        write(sindx,'(I0)') ind(ii)
+        if (len(trim(adjustl(shellNamesTmp(ii)))) + len(trim(sindx)) > sc) then
+          call error("Shell labels are too long: "//trim(adjustl(shellNamesTmp(ii)))//trim(sindx))
+        else
+          shellNamesTmp(ii) = trim(adjustl(shellNamesTmp(ii)))//trim(sindx)
+        end if
+      end if
+    end do
+    deallocate(ind)
+
+  end subroutine getShellNames
+
 
 end module dftbp_orbitals